Post day 5 stuffs
Added Mining tool and name component.

Reorganize the files and put them into 2 folder, 1 for itemtypecomps and
1 for iteminstancecomps.

The purpose is to make it easier to see which components are for which
components.

I also included a g_all.hpp inside items/ folder.

This can be used to include ALL the header in the items folder.

This is mainly used by ItemManager and World.